Update actions footer list view prescriptions
