+++ client/packages/invoices/src/Prescriptions/ListView/Toolbar.tsx
@@ -1,45 +1,17 @@
 import React, { FC } from 'react';
 import {
-  DropdownMenu,
-  DropdownMenuItem,
   useTranslation,
-  DeleteIcon,
   AppBarContentPortal,
   SearchBar,
   FilterController,
   FilterRule,
-  useDeleteConfirmation,
 } from '@openmsupply-client/common';
-import {
-  PrescriptionRowFragment,
-  ListParams,
-  usePrescriptionList,
-} from '../api';
-import { canDeletePrescription } from '../../utils';
+import { PrescriptionRowFragment } from '../api';
 
 export const Toolbar: FC<{
   filter: FilterController;
-  listParams: ListParams;
-}> = ({ filter, listParams }) => {
+}> = ({ filter }) => {
   const t = useTranslation();
-  const {
-    delete: { deletePrescriptions },
-    selectedRows,
-  } = usePrescriptionList(listParams);
-
-  const confirmAndDelete = useDeleteConfirmation({
-    selectedRows,
-    deleteAction: deletePrescriptions,
-    canDelete: selectedRows.every(canDeletePrescription),
-    messages: {
-      confirmMessage: t('messages.confirm-delete-prescriptions', {
-        count: selectedRows.length,
-      }),
-      deleteSuccess: t('messages.deleted-prescriptions', {
-        count: selectedRows.length,
-      }),
-    },
-  });
 
   const key = 'otherPartyName' as keyof PrescriptionRowFragment;
   const filterString =
@@ -61,12 +33,6 @@ export const Toolbar: FC<{
           filter.onChangeStringFilterRule('otherPartyName', 'like', newValue);
         }}
       />
-
-      <DropdownMenu label={t('label.actions')}>
-        <DropdownMenuItem IconComponent={DeleteIcon} onClick={confirmAndDelete}>
-          {t('button.delete-lines')}
-        </DropdownMenuItem>
-      </DropdownMenu>
     </AppBarContentPortal>
   );
 };
